+++ src/repositories/user.repository.js
@@ -54,3 +54,13 @@ export const getAllStoreReviews = async (restaurantId, cursor) => {
 
   return reviews;
 }; 
+
+export const getAllUserReviews = async (userId, cursor) => {
+  const reviews = await prisma.review.findMany({
+    select: { id: true, content: true, restaurantId: true, userId: true },
+    where: { userId: userId, id: { gt: cursor } },
+    orderBy: { id: "asc" },
+    take: 3,
+  });
+  return reviews;
+};
