+++ src/services/review.service.js
@@ -1,6 +1,5 @@
 import { createReview, checkRestaurantExists } from '../repositories/review.repository.js';
-import { getAllUserReviews } from '../repositories/review.repository.js';
-import { responseFromReviews } from '../dtos/review.dto.js';
+
 
 export const addReview = async (data) => {
   const exists = await checkRestaurantExists(data.restaurantId);
@@ -16,7 +15,3 @@ export const addReview = async (data) => {
   
 };
 
-export const listUserReviews = async (userId, cursor = 0) => {
-  const reviews = await getAllUserReviews(userId,cursor);
-  return responseFromReviews(reviews);
-};
